Fixes #40
In view, my  login function name is also login. However there is login(request, user) method in django and when this method is trying to work, because of the same name login view also trying to work. So this causes error.
I fixed this error with changing my login view as login_user.